Factor out text-type.ts.
