+++ src/textbot/base-conversation-handlers.ts
@@ -1,10 +1,10 @@
-import React from "react";
 import {
   ConversationStatus,
   ConversationResponse,
   deserializeConversationState,
   serializeConversationState,
 } from "./conversation";
+import { convertToText, TextType } from "./text-type";
 
 export type BaseConversationState = {
   handlerName: string;
@@ -17,33 +17,6 @@ type ConversationHandlerMethod = () =>
 type ResponseOptions<State extends BaseConversationState> = {
   stateUpdates?: Partial<State>;
 };
-
-type TextType = string | string[] | JSX.Element;
-
-function convertToText(text: TextType): string {
-  if (typeof text === "string") {
-    return text;
-  }
-  if (Array.isArray(text)) {
-    return text.join("\n");
-  } else {
-    const chunks: string[] = [];
-    for (let child of React.Children.toArray(text.props.children)) {
-      if (typeof child === "string") {
-        chunks.push(child);
-      } else if (typeof child === "object" && "key" in child) {
-        if (child.type === "br") {
-          chunks.push("\n");
-        } else {
-          console.log(`Not sure how to render <${child.type.toString()}>.`);
-        }
-      } else {
-        console.log(`Not sure how to render "${child.toString()}".`);
-      }
-    }
-    return chunks.join("");
-  }
-}
 
 export abstract class BaseConversationHandlers<
   State extends BaseConversationState
